+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added: Custom names for Prime 1 elevators
+- Changed: Custom elevator names across both games now used throughout the entire UI
+
 ### Metroid Prime - Patcher Changes
 
 - Added: New Nothing model.